Likelion/pyproject.toml
[project]
name = "likelion"
version = "0.1.0"
description = ""
authors = [
    {name = "luckylee524",email = "dbleecomkr@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "django-allauth (>=65.9.0,<66.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "cryptography (>=45.0.4,<46.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "boto3 (>=1.38.46,<2.0.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)"
]

[tool.poetry]
packages = [{include = "likelion", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
